Adds all calculations for being able to generate animations
Code is now able to generate animation (background and interference layer) with the interference layer being properly scaled based on the ratio of distance from viewer to background and viewer to foreground (parallax ratio).

Code also shows using the parallax calculation to set the X (viewer movement) with a known P (animation period). This stuff will not likely be placed directly in the `Kinegram` class.

Specifics:
* added method for setting pixel width of kine object
* Moved pixel width setting to instantiation of object.
* added setting of parallax ratio with method (`parallaxRatio` is now set correctly)
* modified overlay calculations to properly use parallax value

Addtionally:
* updates README to fully document all calculations
* adds diagram for parallax calculations